fix backend

+++ appointmentManagement/appointmentManagement_backend/appointment_Management/src/main/java/dke/appointment_Management/repository/AppointmentSeriesRepository.java
@@ -3,5 +3,5 @@ package dke.appointment_Management.repository;
 import dke.appointment_Management.entity.AppointmentSeries;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentSeriesRepository extends JpaRepository<AppointmentSeries, Integer> {
+public interface AppointmentSeriesRepository extends JpaRepository<AppointmentSeries, Long> {
 }
